Finish implementation of interactive modify
