+++ src/WingetCreateCLI/PromptHelper.cs
@@ -69,37 +69,15 @@ namespace Microsoft.WingetCreateCLI
         /// <typeparam name="T">Model type.</typeparam>
         /// <param name="model">Instance of object model.</param>
         /// <param name="exitMenuWord">Exit keyword to be shown to the user to exit the navigational menu.</param>
-        public static void PromptPropertiesWithMenu<T>(T model, string exitMenuWord)
+        /// <param name="modelName">If a non-null string is provided, a menu item to display a preview of the model will be added to the selection list</param>
+        public static void PromptPropertiesWithMenu<T>(T model, string exitMenuWord, string modelName = null)
         {
             Console.Clear();
-
-            var properties = model.GetType().GetProperties().ToList();
-            var optionalProperties = properties.Where(p =>
-                p.GetCustomAttribute<RequiredAttribute>() == null &&
-                p.GetCustomAttribute<JsonPropertyAttribute>() != null).ToList();
-
-            var fieldList = properties
-                .Select(property => property.Name)
-                .Where(pName => !NonEditableOptionalFields.Any(field => field == pName) && !NonEditableRequiredFields.Any(field => field == pName)).ToList();
-
-            // Filter out fields if an installerType is present
-            var installerTypeProperty = model.GetType().GetProperty(nameof(InstallerType));
-            if (installerTypeProperty != null)
+            var properties = model.GetType().GetProperties();
+            var fieldList = FilterRestrictedFields(model, properties);
+            if (!string.IsNullOrEmpty(modelName))
             {
-                var installerTypeValue = installerTypeProperty.GetValue(model);
-                if (installerTypeValue != null)
-                {
-                    var installerType = (InstallerType)installerTypeValue;
-                    if (installerType == InstallerType.Msi || installerType == InstallerType.Exe)
-                    {
-                        fieldList.Add(nameof(Installer.ProductCode));
-                    }
-                    else if (installerType == InstallerType.Msix || installerType == InstallerType.Appx)
-                    {
-                        fieldList = fieldList.Where(pName => !MsixExclusionList.Any(field => field == pName)).ToList();
-                        fieldList.AddRange(MsixInclusionList);
-                    }
-                }
+                fieldList.Add(Resources.DisplayPreview_MenuItem);
             }
 
             fieldList.Add(exitMenuWord);
@@ -114,16 +92,61 @@ namespace Microsoft.WingetCreateCLI
                 {
                     break;
                 }
-
-                if (selectedField == nameof(InstallerManifest.Installers))
+                else if (selectedField == Resources.DisplayPreview_MenuItem)
                 {
-                    Commands.NewCommand.DisplayInstallersAsMenuSelection(model as InstallerManifest);
+                    Console.Clear();
+                    Console.WriteLine();
+                    Logger.InfoLocalized(nameof(Resources.DisplayPreviewOfItems), modelName);
+                    Console.WriteLine(Serialization.Serialize(model));
+                    Console.WriteLine();
+                }
+                else if (selectedField == nameof(InstallerManifest.Installers))
+                {
+                    DisplayInstallersAsMenuSelection(model as InstallerManifest);
                 }
                 else
                 {
-
                     var selectedProperty = properties.First(p => p.Name == selectedField);
                     PromptPropertyAndSetValue(model, selectedField, selectedProperty.GetValue(model));
+                }
+            }
+        }
+
+        /// <summary>
+        /// Displays all installers from an Installer manifest as a selection menu.
+        /// </summary>
+        /// <param name="installerManifest">Installer manifest model.</param>
+        public static void DisplayInstallersAsMenuSelection(InstallerManifest installerManifest)
+        {
+            Console.Clear();
+
+            while (true)
+            {
+                List<string> selectionList = GenerateInstallerSelectionList(installerManifest.Installers, out Dictionary<string, Installer> installerSelectionMap);
+                var selectedItem = Prompt.Select(Resources.SelectInstallerToEdit_Message, selectionList);
+
+                if (selectedItem == Resources.SaveAndExit_MenuItem)
+                {
+                    break;
+                }
+                else if (selectedItem == Resources.AllInstallers_MenuItem)
+                {
+                    Installer installerCopy = new Installer();
+                    PromptPropertiesWithMenu(installerCopy, Resources.SaveAndExit_MenuItem);
+                    ApplyChangesToIndividualInstallers(installerCopy, installerManifest.Installers);
+                }
+                else if (selectedItem == Resources.DisplayPreview_MenuItem)
+                {
+                    Console.Clear();
+                    Console.WriteLine();
+                    Logger.InfoLocalized(nameof(Resources.DisplayPreviewOfSelectedInstaller_Message));
+                    Console.WriteLine(Serialization.Serialize(installerManifest));
+                    Console.WriteLine();
+                }
+                else
+                {
+                    Installer selectedInstaller = installerSelectionMap[selectedItem];
+                    PromptPropertiesWithMenu(selectedInstaller, Resources.SaveAndExit_MenuItem, selectedItem.Split(':')[0]);
                 }
             }
         }
@@ -286,6 +309,119 @@ namespace Microsoft.WingetCreateCLI
             }
         }
 
+        private static List<string> FilterRestrictedFields<T>(T model, PropertyInfo[] properties)
+        {
+            var fieldList = properties
+                .Select(property => property.Name)
+                .Where(pName =>
+                    !NonEditableOptionalFields.Any(field => field == pName) &&
+                    !NonEditableRequiredFields.Any(field => field == pName)).ToList();
+
+            // Filter out fields if an installerType is present
+            var installerTypeProperty = model.GetType().GetProperty(nameof(InstallerType));
+            if (installerTypeProperty != null)
+            {
+                var installerTypeValue = installerTypeProperty.GetValue(model);
+                if (installerTypeValue != null)
+                {
+                    var installerType = (InstallerType)installerTypeValue;
+                    if (installerType == InstallerType.Msi || installerType == InstallerType.Exe)
+                    {
+                        fieldList.Add(nameof(Installer.ProductCode));
+                    }
+                    else if (installerType == InstallerType.Msix || installerType == InstallerType.Appx)
+                    {
+                        fieldList = fieldList.Where(pName => !MsixExclusionList.Any(field => field == pName)).ToList();
+                        fieldList.AddRange(MsixInclusionList);
+                    }
+                }
+            }
+
+            return fieldList;
+        }
+
+        private static List<string> GenerateInstallerSelectionList(List<Installer> installers, out Dictionary<string, Installer> installerSelectionMap)
+        {
+            installerSelectionMap = new Dictionary<string, Installer>();
+            int index = 1;
+            foreach (Installer installer in installers)
+            {
+                var installerTuple = string.Join(" | ", new[]
+                    {
+                        installer.Architecture.ToEnumAttributeValue(),
+                        installer.InstallerType.ToEnumAttributeValue(),
+                        installer.Scope?.ToEnumAttributeValue(),
+                        installer.InstallerLocale,
+                        installer.InstallerUrl,
+                    }.Where(s => !string.IsNullOrEmpty(s)));
+
+                var installerMenuItem = string.Format(Resources.InstallerSelection_MenuItem, index, installerTuple);
+                installerSelectionMap.Add(installerMenuItem, installer);
+                index++;
+            }
+
+            List<string> selectionList = new List<string>() { Resources.AllInstallers_MenuItem };
+            selectionList.AddRange(installerSelectionMap.Keys);
+            selectionList.AddRange(new[] { Resources.DisplayPreview_MenuItem, Resources.SaveAndExit_MenuItem });
+            return selectionList;
+        }
+
+        private static void ApplyChangesToIndividualInstallers(Installer installerCopy, List<Installer> installers)
+        {
+            // Skip architecture as the default value when instantiated is x86.
+            var modifiedFields = installerCopy.GetType().GetProperties()
+                .Select(prop => prop)
+                .Where(pi =>
+                    pi.GetValue(installerCopy) != null &&
+                    pi.Name != nameof(Installer.Architecture) &&
+                    pi.Name != nameof(Installer.AdditionalProperties));
+
+            foreach (var field in modifiedFields)
+            {
+                foreach (Installer installer in installers)
+                {
+                    var fieldValue = field.GetValue(installerCopy);
+                    var prop = installer.GetType().GetProperty(field.Name);
+                    if (prop.PropertyType.IsValueType)
+                    {
+                        prop.SetValue(installer, fieldValue);
+                    }
+                    else if (fieldValue is IList list)
+                    {
+                        prop.SetValue(installer, list.DeepClone());
+                    }
+                    else if (fieldValue is Dependencies dependencies)
+                    {
+                        ApplyDependencyChangesToInstaller(dependencies, installer);
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Clones any non-null property values of the dependencies object and assigns them to the provided installer object.
+        /// </summary>
+        /// <param name="dependencies">Dependencies object with new values.</param>
+        /// <param name="installer">Installer object to assign new changes to.</param>
+        private static void ApplyDependencyChangesToInstaller(Dependencies dependencies, Installer installer)
+        {
+            var modifiedFields = dependencies.GetType().GetProperties()
+                .Select(prop => prop)
+                .Where(pi => pi.GetValue(dependencies) != null);
+
+            foreach (var field in modifiedFields.Where(f => f.Name != nameof(Installer.AdditionalProperties)))
+            {
+                var fieldValue = field.GetValue(dependencies);
+                installer.Dependencies ??= new Dependencies();
+                var prop = installer.Dependencies.GetType().GetProperty(field.Name);
+
+                if (fieldValue is IList list)
+                {
+                    prop.SetValue(installer.Dependencies, list.DeepClone());
+                }
+            }
+        }
+
         private static void PromptEnum(object model, PropertyInfo property, Type enumType, bool required, string message, string defaultValue = null)
         {
             var enumList = Enum.GetNames(enumType);
@@ -305,7 +441,6 @@ namespace Microsoft.WingetCreateCLI
         private static void PromptForItemList<T>(List<T> items)
             where T : new()
         {
-            var serializer = Serialization.CreateSerializer();
             string selection = string.Empty;
             while (selection != Resources.Back_MenuItem)
             {
@@ -313,7 +448,7 @@ namespace Microsoft.WingetCreateCLI
                 if (items.Any())
                 {
                     Logger.InfoLocalized(nameof(Resources.DisplayPreviewOfItems), typeof(T).Name);
-                    string serializedString = serializer.Serialize(items);
+                    string serializedString = Serialization.Serialize(items);
                     Console.WriteLine(serializedString);
                     Console.WriteLine();
                 }
@@ -345,7 +480,7 @@ namespace Microsoft.WingetCreateCLI
 
         private static void PromptSubfieldProperties<T>(T field, PropertyInfo property, object model)
         {
-            PromptPropertiesWithMenu(field, Resources.None_MenuItem);
+            PromptPropertiesWithMenu(field, Resources.SaveAndExit_MenuItem);
             if (field.IsEmptyObject())
             {
                 property.SetValue(model, null);
